break function_app into modules
